### What
Here we do two main things:
1. Add more testing for Vamana helper functions
2. Add an option to `best_first_O4()` to skip computing top_k ids and distances. This is useful in Vamana ingestion where they are unused.

### Performance
The accuracy numbers are the same and ingestion is sometimes faster, but sometimes not, so likely just variability in my local machine is causing the change.

**Charts**
Before

![ingestion_time_vs_accuracy](https://github.com/user-attachments/assets/efc4c8ed-6487-46d7-a644-a7b7a623e640)

After

![ingestion_time_vs_accuracy](https://github.com/user-attachments/assets/263dcd9d-f752-4091-9a21-0efff002f211)

**Raw numbers**
Before
```
VAMANA_l_build=40_r_max_degree=10
  Ingestion (count: 1):
    Average Time: 181.4188 seconds
  Query (count: 5):
    Average Time: 2.4649 seconds
    Average Accuracy: 0.7436

VAMANA_l_build=40_r_max_degree=15
  Ingestion (count: 1):
    Average Time: 260.3602 seconds
  Query (count: 5):
    Average Time: 1.9147 seconds
    Average Accuracy: 0.8676

VAMANA_l_build=40_r_max_degree=20
  Ingestion (count: 1):
    Average Time: 341.2821 seconds
  Query (count: 5):
    Average Time: 2.0450 seconds
    Average Accuracy: 0.9201

VAMANA_l_build=40_r_max_degree=25
  Ingestion (count: 1):
    Average Time: 419.1305 seconds
  Query (count: 5):
    Average Time: 1.9463 seconds
    Average Accuracy: 0.9439

VAMANA_l_build=40_r_max_degree=30
  Ingestion (count: 1):
    Average Time: 456.7194 seconds
  Query (count: 5):
    Average Time: 2.0422 seconds
    Average Accuracy: 0.9562

VAMANA_l_build=40_r_max_degree=35
  Ingestion (count: 1):
    Average Time: 480.5166 seconds
  Query (count: 5):
    Average Time: 2.0756 seconds
    Average Accuracy: 0.9624

VAMANA_l_build=40_r_max_degree=40
  Ingestion (count: 1):
    Average Time: 503.7234 seconds
  Query (count: 5):
    Average Time: 2.1702 seconds
    Average Accuracy: 0.9662
```

After
```
VAMANA_l_build=40_r_max_degree=10
  Ingestion (count: 1):
    Average Time: 157.6301 seconds
  Query (count: 5):
    Average Time: 1.7090 seconds
    Average Accuracy: 0.7436

VAMANA_l_build=40_r_max_degree=15
  Ingestion (count: 1):
    Average Time: 239.2754 seconds
  Query (count: 5):
    Average Time: 1.9881 seconds
    Average Accuracy: 0.8676

VAMANA_l_build=40_r_max_degree=20
  Ingestion (count: 1):
    Average Time: 330.6436 seconds
  Query (count: 5):
    Average Time: 2.3247 seconds
    Average Accuracy: 0.9201

VAMANA_l_build=40_r_max_degree=25
  Ingestion (count: 1):
    Average Time: 410.9352 seconds
  Query (count: 5):
    Average Time: 2.1819 seconds
    Average Accuracy: 0.9439

VAMANA_l_build=40_r_max_degree=30
  Ingestion (count: 1):
    Average Time: 463.5973 seconds
  Query (count: 5):
    Average Time: 2.4142 seconds
    Average Accuracy: 0.9562

VAMANA_l_build=40_r_max_degree=35
  Ingestion (count: 1):
    Average Time: 500.3645 seconds
  Query (count: 5):
    Average Time: 2.4888 seconds
    Average Accuracy: 0.9624

VAMANA_l_build=40_r_max_degree=40
  Ingestion (count: 1):
    Average Time: 521.9184 seconds
  Query (count: 5):
    Average Time: 2.6005 seconds
    Average Accuracy: 0.9662
```

### Testing
Tests pass.

### Benchmarking
I also tested `best_first_O4` vs `best_first_O5` (with `std::cout << _timing_data.dump();`), and `best_first_O4` (which we use) seems to be faster:
```
auto best_first_O4 count: 5, sum: 20.04 µs, avg: 4.01 µs
auto best_first_O5: count: 5, sum: 40.71 µs, avg: 8.14 µs
```