@jparismorgan 2024-08-07
> I think the names of the functions here are really not descriptive. Can we avoid the O1-4 notation and use better names or at least document the difference between the different best first implementations?

I do want to do this, but would prefer to do it in a future refactor, likely alongside fixing `best_first_O2` and `best_first_O3`, which don't compile. And we can also do it for the functions in `src/include/detail/graph/greedy_search.h`. Is that okay with you?